+++ pbdoom/frames/color_settings.h
@@ -1,0 +1,8 @@
+#ifndef PBDOOM_FRAMES_COLOR_SETTINGS_H
+#define PBDOOM_FRAMES_COLOR_SETTINGS_H
+
+#include "frame.h"
+
+extern frame color_settings_frame;
+
+#endif /* PBDOOM_FRAMES_COLOR_SETTINGS_H */