+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - DPI (dp) button sizes for uniform sizing across devices [#2](https://github.com/imustafin/pbdoom/issues/2)
 - Added vertical margins for game display (refactor display elements into frames)
+- Added color settings menu
 
 ### Fixed